serving_pipeline/frontend/requirements.txt
gradio==5.23.3
requests==2.32.3
pillow==11.1.0
python-dotenv==1.0.0
pandas==2.2.3
plotly==5.24.1
